ファイルを分離
