+++ graph_base.hpp
@@ -11,6 +11,15 @@ struct edge {
 	// 重み付き辺の追加
 	edge(int _src, int _to, T _cost) : src(_src), to(_to), cost(_cost) {}
 };
+//edgeに関する演算子
+template <typename T>
+bool operator<(const edge<T>& lhs, const edge<T>& rhs) {
+	return lhs.cost < rhs.cost;
+}
+template <typename T>
+bool operator>(const edge<T>& lhs, const edge<T>& rhs) {
+	return lhs.cost > rhs.cost;
+}
 // 辺の集合
 template <typename T>
 using Edges = vector<edge<T>>;
